Move API request hacks to a separate file
